requirements.txt
# PDF Processing
pdfminer.six==20221105

# Deep Learning
torch==2.1.0+cu121  # phù hợp với CUDA 12.1
transformers==4.35.0
datasets==2.14.6
accelerate==0.25.0

# Data Processing
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Progress Bar
tqdm==4.66.1

# Streamlit
streamlit==1.29.0

# Utilities
sentencepiece==0.1.99  # cần thiết cho xlm-roberta
protobuf==3.20.3 # để tránh lỗi tương thích với transformers